feat(registry-dash): move timeframe selector into global filter panel
The timeframe selector (6h–24m) was per-tab, resetting when switching
between Domain Activity, Financials, Forecasting, etc. Now it lives in
the shared filter panel as a global signal, persisting across all tabs.
Overview also gains dynamic timeframe awareness instead of hardcoded
12-month lookback.
