+++ console-webapp/src/app/registry-dash/filter-panel/filter-panel.component.scss
@@ -79,6 +79,11 @@
   border-top: 1px solid #f0f0f0;
 }
 
+.time-range {
+  width: 100%;
+  padding-top: 12px;
+}
+
 .filter-select {
   width: 220px;
